Allow skip gorse

+++ app/services/gorse_service.rb
@@ -25,6 +25,7 @@ class GorseService
     end
 
     def sync_user(user)
+      return if ENV["SKIP_GORSE"]
       response = with_retry do
         connection.post("/api/user") do |req|
           req.headers["X-API-KEY"] = api_key
@@ -41,6 +42,7 @@ class GorseService
     end
 
     def delete_user(user)
+      return if ENV["SKIP_GORSE"]
       response = with_retry do
         connection.delete("/api/user/#{CGI.escape(user.id.to_s)}") do |req|
           req.headers["X-API-KEY"] = api_key
@@ -51,6 +53,7 @@ class GorseService
     end
 
     def sync_item(item)
+      return if ENV["SKIP_GORSE"]
       case item
       when Project
         sync_project_item(item)
@@ -62,6 +65,7 @@ class GorseService
     end
 
     def delete_item(item)
+      return if ENV["SKIP_GORSE"]
       case item
       when Project
         delete_project_item(item)
@@ -73,6 +77,7 @@ class GorseService
     end
 
     def delete_item_by_id(item_id)
+      return if ENV["SKIP_GORSE"]
       response = with_retry do
         connection.delete("/api/item/#{CGI.escape(item_id.to_s)}") do |req|
           req.headers["X-API-KEY"] = api_key
@@ -83,6 +88,7 @@ class GorseService
     end
 
     def sync_feedback(feedback_type, user_id, item, timestamp)
+      return if ENV["SKIP_GORSE"]
       events = []
 
       prefixed_id = case item
@@ -119,6 +125,7 @@ class GorseService
     end
 
     def delete_feedback(feedback_type, user_id, item_id)
+      return if ENV["SKIP_GORSE"]
       response = with_retry do
         connection.delete("/api/feedback/#{CGI.escape(feedback_type.to_s)}/#{CGI.escape(user_id.to_s)}/#{CGI.escape(item_id.to_s)}") do |req|
           req.headers["X-API-KEY"] = api_key
@@ -129,6 +136,7 @@ class GorseService
     end
 
     def get_user_recommendation(user_id, page = 1, per_page = 21, type: :project)
+      return [] if ENV["SKIP_GORSE"]
       offset = (page - 1) * per_page
 
       response = with_retry do
@@ -150,6 +158,7 @@ class GorseService
     end
 
     def get_popular_items(page = 1, per_page = 21, type: :project)
+      return [] if ENV["SKIP_GORSE"]
       offset = (page - 1) * per_page
       category = type == :entry ? "entry" : "project"
